requirements.txt
# =============================================================================
# Mission9 - Python Dependencies
# PySpark + PyTorch GPU + MLflow
# =============================================================================

# -----------------------------------------------------------------------------
# PyTorch with CUDA 12.4 support
# -----------------------------------------------------------------------------
--extra-index-url https://download.pytorch.org/whl/cu124
torch>=2.5.0
torchvision>=0.20.0

# -----------------------------------------------------------------------------
# Apache Spark
# -----------------------------------------------------------------------------
pyspark==3.5.3
pyarrow>=14.0.0

# -----------------------------------------------------------------------------
# Data Processing
# -----------------------------------------------------------------------------
pandas>=2.0.0
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0

# -----------------------------------------------------------------------------
# MLflow Tracking
# -----------------------------------------------------------------------------
mlflow>=2.18.0

# -----------------------------------------------------------------------------
# Jupyter Environment
# -----------------------------------------------------------------------------
jupyterlab>=4.0.0
ipywidgets>=8.0.0
tqdm>=4.66.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0
requests>=2.31.0
